+++ app/lib/routing/work-screen.ts
@@ -23,6 +23,5 @@ export enum WorkScreen {
   tool_meta_editor = "tool_meta_editor",
   tool_batch_meta_editor = "tool_batch_meta_editor",
   tool_data_mapper = "tool_data_mapper",
-  scene_upload_steps_final = "scene_upload_steps_final",
   signin = "signin",
 }