+++ CHANGELOG.md
@@ -28,6 +28,13 @@
 - [Flutter] Poligon Node support with XImage (svg)
 - [Lint] Primal naming & grouping linting for better code export quality. this is tracked sperately on [lint](https://github.com/bridgedxyz/lint)
 
+## [2022.4.0.2] - 2022-04-28
+
+- PR: https://github.com/gridaco/assistant/pull/197
+- Add “Open in Grida” action in design/preview screen like as in code/preview screen.
+- update design preview screen to show preview in a more isolated way - (hide tab bars while previewing)
+- update filekey input to be more visible
+
 ## [2022.4.0.1] - 2022-04-08
 
 - D2C: CSS Duplication reduction